# pkg/api/openapi/gamehub.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GameHub API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/api/v1/metadata/metaobject-definitions": {
      "post": {
        "operationId": "createMetaobjectDefinition",
        "tags": [
          "metadata"
        ],
        "summary": "Create a metaobject definition",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/ContentTypeHeader"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/MetaobjectDefinition"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/MetaobjectDefinitionCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "operationId": "listMetaobjectDefinitions",
        "tags": [
          "metadata"
        ],
        "summary": "List metaobject definitions",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/MetaobjectTypeQuery"
          },
          {
            "$ref": "#/components/parameters/ActiveQuery"
          },
          {
            "$ref": "#/components/parameters/PageSizeQuery"
          },
          {
            "$ref": "#/components/parameters/PageTokenQuery"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MetaobjectDefinitionList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metadata/metaobject-definitions/{definition_id}": {
      "get": {
        "operationId": "getMetaobjectDefinition",
        "tags": [
          "metadata"
        ],
        "summary": "Get one metaobject definition",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MetaobjectDefinition"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch": {
        "operationId": "updateMetaobjectDefinition",
        "tags": [
          "metadata"
        ],
        "summary": "Update one metaobject definition",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/ContentTypeHeader"
          },
          {
            "$ref": "#/components/parameters/IfMatchHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/MetaobjectDefinitionUpdate"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MetaobjectDefinition"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "archiveMetaobjectDefinition",
        "tags": [
          "metadata"
        ],
        "summary": "Archive one metaobject definition",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/IfMatchHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metadata/metaobject-definitions/{definition_id}/entries": {
      "post": {
        "operationId": "createMetaobjectEntry",
        "tags": [
          "metadata"
        ],
        "summary": "Create a metaobject entry",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/ContentTypeHeader"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/MetaobjectEntry"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/MetaobjectEntryCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "operationId": "listMetaobjectEntries",
        "tags": [
          "metadata"
        ],
        "summary": "List metaobject entries",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/PageSizeQuery"
          },
          {
            "$ref": "#/components/parameters/PageTokenQuery"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MetaobjectEntryList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metadata/metaobject-definitions/{definition_id}/entries/{entry_id}": {
      "get": {
        "operationId": "getMetaobjectEntry",
        "tags": [
          "metadata"
        ],
        "summary": "Get one metaobject entry",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/EntryIdPath"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MetaobjectEntry"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch": {
        "operationId": "updateMetaobjectEntry",
        "tags": [
          "metadata"
        ],
        "summary": "Update one metaobject entry",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/EntryIdPath"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          },
          {
            "$ref": "#/components/parameters/ContentTypeHeader"
          },
          {
            "$ref": "#/components/parameters/IfMatchHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/MetaobjectEntryUpdate"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MetaobjectEntry"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteMetaobjectEntry",
        "tags": [
          "metadata"
        ],
        "summary": "Delete one metaobject entry",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/DefinitionIdPath"
          },
          {
            "$ref": "#/components/parameters/EntryIdPath"
          },
          {
            "$ref": "#/components/parameters/IfMatchHeader"
          },
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "headers": {
...
      "ETag": {
        "description": "Numeric resource version wrapped as an entity tag.",
        "schema": {
          "type": "string",
          "example": "\"1\""
        }
      },
      "RateLimitLimit": {
        "description": "Maximum requests allowed in the current window.",
        "schema": {
          "type": "integer"
        }
      },
      "RateLimitRemaining": {
        "description": "Remaining requests in the current window.",
        "schema": {
          "type": "integer"
        }
      },
      "RateLimitReset": {
        "description": "Unix timestamp when the current limit window resets.",
        "schema": {
          "type": "integer"
        }
      },
      "RequestId": {
        "description": "Unique identifier for this request.",
        "schema": {
          "type": "string"
        }
      },
      "RetryAfter": {
        "description": "Seconds to wait before retrying.",
        "schema": {
          "type": "integer"
        }
      }
    },
    "parameters": {
      "AcceptHeader": {
        "name": "Accept",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "default": "application/json"
        }
      },
      "ContentTypeHeader": {
        "name": "Content-Type",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "const": "application/json"
        }
      },
      "CorrelationIdHeader": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "RequestIdHeader": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "IdempotencyKeyHeader": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "IfMatchHeader": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "example": "\"1\""
        }
      },
...
      "DefinitionIdPath": {
        "name": "definition_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "EntryIdPath": {
        "name": "entry_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "MetaobjectTypeQuery": {
        "name": "type",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/MachineKey"
        }
      },
      "ActiveQuery": {
        "name": "active",
        "in": "query",
        "required": false,
        "schema": {
          "type": "boolean"
        }
      },
...
      "PageSizeQuery": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "PageTokenQuery": {
        "name": "page_token",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "requestBodies": {
...
      "MetaobjectDefinition": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectDefinitionRequest"
            }
          }
        }
      },
      "MetaobjectDefinitionUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectDefinitionUpdateRequest"
            }
          }
        }
      },
      "MetaobjectEntry": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectEntryRequest"
            }
          }
        }
      },
      "MetaobjectEntryUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectEntryUpdateRequest"
            }
          }
        }
      }
    },
    "responses": {
      "NoContent": {
        "description": "No content",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          },
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          },
          "RateLimit-Limit": {
            "$ref": "#/components/headers/RateLimitLimit"
          },
          "RateLimit-Remaining": {
            "$ref": "#/components/headers/RateLimitRemaining"
          },
          "RateLimit-Reset": {
            "$ref": "#/components/headers/RateLimitReset"
          }
        }
      },
...
      "MetaobjectDefinitionCreated": {
        "description": "Metaobject definition created",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectDefinition"
            }
          }
        }
      },
      "MetaobjectDefinition": {
        "description": "Metaobject definition",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectDefinition"
            }
          }
        }
      },
      "MetaobjectDefinitionList": {
        "description": "Metaobject definition list",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectDefinitionList"
            }
          }
        }
      },
      "MetaobjectEntryCreated": {
        "description": "Metaobject entry created",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectEntry"
            }
          }
        }
      },
      "MetaobjectEntry": {
        "description": "Metaobject entry",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectEntry"
            }
          }
        }
      },
      "MetaobjectEntryList": {
        "description": "Metaobject entry list",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MetaobjectEntryList"
            }
          }
        }
      },
      "BadRequest": {
        "$ref": "#/components/responses/Problem400"
      },
      "UnsupportedMediaType": {
        "$ref": "#/components/responses/Problem415"
      },
      "NotAcceptable": {
        "$ref": "#/components/responses/Problem406"
      },
      "NotFound": {
        "$ref": "#/components/responses/Problem404"
      },
      "Conflict": {
        "$ref": "#/components/responses/Problem409"
      },
      "PreconditionFailed": {
        "$ref": "#/components/responses/Problem412"
      },
      "ValidationFailed": {
        "$ref": "#/components/responses/Problem422"
      },
      "RateLimited": {
        "description": "Rate limited",
        "headers": {
          "RateLimit-Limit": {
            "$ref": "#/components/headers/RateLimitLimit"
          },
          "RateLimit-Remaining": {
            "$ref": "#/components/headers/RateLimitRemaining"
          },
          "RateLimit-Reset": {
            "$ref": "#/components/headers/RateLimitReset"
          },
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "InternalError": {
        "$ref": "#/components/responses/Problem500"
      },
      "Problem400": {
        "description": "Bad request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Problem406": {
        "description": "Not acceptable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem409": {
        "description": "Conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Problem415": {
        "description": "Unsupported media type",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem422": {
        "description": "Validation failed",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem500": {
        "description": "Internal server error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "MetaobjectDefinitionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "name",
          "fields"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/MachineKey"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldDefinition"
            }
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "MetaobjectDefinition": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MetaobjectDefinitionRequest"
          },
          {
            "$ref": "#/components/schemas/ResourceFields"
          }
        ]
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      }
    }
  }
}